+++ ui/ui-frontend/src/components/ProductsInfo/AccessoriesProductsByTpeInfo.jsx
@@ -44,9 +44,11 @@ const AccessoriesProductsByTpeInfo = () => {
     if (localStorage.getItem('token')) {
       const datad = await CartService.getProductCategoryAndProductId(category, accessoryId).then()
         .catch((err) => {
-          console.log(err)
-          navigate("/login")
-          localStorage.clear();
+          if (err.response.status === 401) {
+            console.log(err.response.data)
+            navigate("/login")
+            localStorage.clear();
+        }
         });
 
       console.log(datad.data);
@@ -73,9 +75,11 @@ const AccessoriesProductsByTpeInfo = () => {
             console.log(response);
             alert("Item added successfully");
           }).catch((err) => {
-            console.log(err.response.data)
-            navigate("/login")
-            localStorage.clear();
+            if (err.response.status === 401) {
+              console.log(err.response.data)
+              navigate("/login")
+              localStorage.clear();
+          }
           });
         } else {
           alert(" products Left");
@@ -86,9 +90,11 @@ const AccessoriesProductsByTpeInfo = () => {
         await CartService.updateQuantity(datad.data.cartId, username, qty).then(() => {
           alert("Cart contains " + qty + " " + datad.data.productName);
         }).catch((err) => {
-          console.log(err.response.data)
-          navigate("/login")
-          localStorage.clear();
+          if (err.response.status === 401) {
+            console.log(err.response.data)
+            navigate("/login")
+            localStorage.clear();
+        }
         });
       }
     }
