+++ ui/ui-frontend/src/components/ProductsInfo/FootwearProductsByTpeInfo.jsx
@@ -41,9 +41,11 @@ const FootwearProductsByTypeInfo = () => {
     if (localStorage.getItem('token')) {
       const datad = await CartService.getProductCategoryAndProductId(category, footWearId).then()
         .catch((err) => {
-          console.log(err.response.data)
-          navigate("/login")
-          localStorage.clear();
+          if (err.response.status === 401) {
+            console.log(err.response.data)
+            navigate("/login")
+            localStorage.clear();
+        }
         });
 
       console.log(datad.data);
@@ -68,9 +70,11 @@ const FootwearProductsByTypeInfo = () => {
             //   console.log(response);
             alert("Item added successfully");
           }).catch((err) => {
-            console.log(err.response.data)
-            navigate("/login")
-            localStorage.clear();
+            if (err.response.status === 401) {
+              console.log(err.response.data)
+              navigate("/login")
+              localStorage.clear();
+          }
           });
         } else {
           alert(" products Left");
@@ -81,9 +85,11 @@ const FootwearProductsByTypeInfo = () => {
         await CartService.updateQuantity(datad.data.cartId, username, qty).then(() => {
           alert("Cart contains " + qty + " " + datad.data.productName);
         }).catch((err) => {
-          console.log(err.response.data)
-          navigate("/login")
-          localStorage.clear();
+          if (err.response.status === 401) {
+            console.log(err.response.data)
+            navigate("/login")
+            localStorage.clear();
+        }
         });
       }
     }
